Implement password strength meter
Define prop showMeter to enable the strength meter. Implement the meter
as a progress bar with a max of 100. Define passwordStrength function,
which is called in validationError which then sets the strength in the
state. Define pwStrengthClass function, which determines what class
should be applied to the meter, which is then given a color based on the
class.