Add keyboard & mouse transformation controls
Create ObjectTransformer to provide keyboard controls to change the
transformation options and a function to transform a ThreeDObject. Use
ObjectTransformer in ThreeDObject to implement the event to set the
transformation on the mousemove event. Create DeviceInput as a wrapper
for mouse and keyboard events which has the main purpose of not
capturing keyboard auto-repeat signals for the keydown events. Provide a
callback for transformation details updates via the ThreeDObject's
onTransformDetailsUpdate prop, which can be used to update a UI widget
with the current transformation mode and last selected ThreeDObject's
property values.