Implement ArrayBuilder for the property editor
Define ArrayBuilder type for the PropEditor, which is used to build a
named array of objects with a specific definition.

-----

Unlike the ArrayEditor which loads data from the database, the
ArrayBuilder is used to create custom objects from a defined format. For
example, to create an array of objects that each have the properties
label, url and color with a maximum number of 5 elements, it can be
specified with the following in a component's propTypes:

exampleArr: {
    type: 'array',
    maxRows: 5,
    fields {label: 'string', url: 'string', color: 'color'}
}

You can also specify the defaultProps for that array, for example:

exampleArr: [{label: 'No label', url: '', color: '#ff0000'}]

This will result in a single row in the PropEditor with those fields. If
multiple objects are specified in the defaultProps array, then the array
will be initially populated with those values but will use the first
element as a placeholder for new rows added.

A '+' button is provided for these arrays in the PropEditor to add
elements, and each row has a '-' to remove that element.