Fix CurvedModel video rendering
Re-add 'display: none;' to the rendered CurvedModel element (as it was
originally done in the ThreeDCurvedImage model) so that videos display
correctly. Use a React ref for the video element (instead of getting it
by ID) when creating and assigning the video texture so that multiple
ThreeDObjects in a scene that are using the CurvedModel render strategy
don't share the same video.